+++ src/components/CustomRoutes/CustomRoutes.js
@@ -2,8 +2,8 @@ import React from 'react';
 import Login from '../../pages/Login';
 import Home from '../../pages/Home/Home';
 import ErrorPage from '../../pages/ErrorPage';
-import CabRequest from '../CabRequest/CabRequest';
-import EmployeeCabRequest from '../CabRequest/EmployeeCabRequest';
+import CabRequest from '../CabRequest/Admin/CabRequest';
+import EmployeeCabRequest from '../CabRequest/Employee/EmployeeCabRequest';
 import CabRequestForm from '../../pages/CabRequestForm/CabRequestForm';
 import { Route, Routes, Navigate } from 'react-router-dom';
 import { AuthContext } from '../../context/AuthContext';
