+++ src/services/CabRequestService.ts
@@ -34,6 +34,13 @@ class CabRequestService {
       { status: 'APPROVED', vendorId: vendorId }
     );
   }
+
+  static declineRequest(requestId: number) {
+    return axios.put(
+      `https://shuttle-service-tw.vercel.app/api/v1/cab-request/${requestId}`,
+      { status: 'DECLINED' }
+    );
+  }
 }
 
 export default CabRequestService;
